requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==7.0.0
idna==3.6
influxdb-client==1.41.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
